规避探针增强MonitorCollector后第二个new Runnable()偶尔会变成Unavailable Anonymous Inner Class!的问题
